Revert "Adjust emoji offsets (#28635)"
This reverts commit d9c0c4974d26d38a48b1fb8bd0c9cc07fa80184f.
