+++ shared/chat/conversation/input-area/suggestors/emoji.tsx
@@ -2,7 +2,13 @@ import * as C from '@/constants'
 import * as Common from './common'
 import * as Kb from '@/common-adapters'
 import * as React from 'react'
-import {type EmojiData, RPCToEmojiData, emojiData} from '@/common-adapters/emoji'
+import {
+  emojiSearch,
+  emojiDataToRenderableEmoji,
+  renderEmoji,
+  type EmojiData,
+  RPCToEmojiData,
+} from '@/util/emoji'
 
 export const transformer = (
   emoji: EmojiData,
@@ -27,7 +33,7 @@ const ItemRenderer = (p: Common.ItemRendererProps<EmojiData>) => {
       ])}
       gap="small"
     >
-      <Kb.Emoji emojiData={item} showTooltip={false} size={24} />
+      {renderEmoji({emoji: emojiDataToRenderableEmoji(item), showTooltip: false, size: 24})}
       <Kb.Text type="BodySmallSemibold">{item.short_name}</Kb.Text>
     </Kb.Box2>
   )
@@ -55,17 +61,17 @@ const useDataSource = (filter: string) => {
   }
 
   // prefill data with stock emoji
-  let results: Array<EmojiData> = emojiData.emojiSearch(filter, 50)
+  let emojiData: Array<EmojiData> = emojiSearch(filter, 50)
 
   if (userEmojis) {
     const userEmoji = userEmojis
       .filter(emoji => emoji.alias.toLowerCase().includes(filter))
       .map(emoji => RPCToEmojiData(emoji, false))
-    results = userEmoji.sort((a, b) => a.short_name.localeCompare(b.short_name)).concat(results)
+    emojiData = userEmoji.sort((a, b) => a.short_name.localeCompare(b.short_name)).concat(emojiData)
   }
 
   return {
-    items: results,
+    items: emojiData,
     loading: userEmojisLoading,
   }
 }
