+++ shared/chat/conversation/messages/emoji-row.tsx
@@ -4,6 +4,7 @@ import {useOrdinal} from './ids-context'
 import * as Kb from '@/common-adapters'
 import type * as T from '@/constants/types'
 import {EmojiPickerDesktop} from '@/chat/emoji-picker/container'
+import {renderEmoji, RPCUserReacjiToRenderableEmoji} from '@/util/emoji'
 
 type OwnProps = {
   className?: string
@@ -120,14 +121,13 @@ const HoverEmoji = (props: {emoji: T.RPCGen.UserReacji; onClick: () => void}) =>
       hoverColor={Kb.Styles.globalColors.transparent}
       style={styles.emojiBox}
     >
-      <Kb.Emoji
-        userReacji={props.emoji}
-        noAnim={!hovering}
-        showTooltip={false}
-        size={hovering ? 22 : 18}
-        style={styles.hoverEmoji}
-        virtualText={true}
-      />
+      {renderEmoji({
+        emoji: RPCUserReacjiToRenderableEmoji(props.emoji, !hovering),
+        showTooltip: false,
+        size: hovering ? 22 : 18,
+        style: styles.hoverEmoji,
+        virtualText: true,
+      })}
     </Kb.ClickableBox>
   )
 }
