+++ shared/common-adapters/emoji.d.ts
