+++ shared/common-adapters/emoji.native.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react'
 import * as Styles from '@/styles'
 import {isAndroid} from '@/constants/platform'
-import {emojiIndexByName} from '../markdown/emoji-gen'
-import Text from '../text'
+import {emojiIndexByName} from './markdown/emoji-gen'
+import Text from './text'
 
-import type {Props} from './native-emoji'
+import type {Props} from './emoji'
 
 const familyOverride = isAndroid ? {fontFamily: ''} : {}
