+++ shared/common-adapters/index.d.ts
@@ -33,7 +33,6 @@ export {default as CopyableText} from './copyable-text'
 export {default as Divider} from './divider'
 export {default as Dropdown, DropdownButton, InlineDropdown} from './dropdown'
 export {default as DragAndDrop} from './drag-and-drop'
-export {default as NativeEmoji} from './emoji/native-emoji'
 export {default as Emoji} from './emoji'
 export {default as ErrorBoundary} from './error-boundary'
 export {default as FloatingBox} from './floating-box'
