--- shared/ios/Podfile.properties.json
@@ -1,3 +1,0 @@
-{
-  "ios.buildReactNativeFromSource": "true"
-}
