--- shared/patches/react-native+0.81.4.patch
@@ -1,20 +1,0 @@
-diff --git a/node_modules/react-native/ReactCommon/react/renderer/textlayoutmanager/platform/ios/react/renderer/textlayoutmanager/RCTTextLayoutManager.mm b/node_modules/react-native/ReactCommon/react/renderer/textlayoutmanager/platform/ios/react/renderer/textlayoutmanager/RCTTextLayoutManager.mm
-index 216bb23..ec0578d 100644
---- a/node_modules/react-native/ReactCommon/react/renderer/textlayoutmanager/platform/ios/react/renderer/textlayoutmanager/RCTTextLayoutManager.mm
-+++ b/node_modules/react-native/ReactCommon/react/renderer/textlayoutmanager/platform/ios/react/renderer/textlayoutmanager/RCTTextLayoutManager.mm
-@@ -411,10 +411,11 @@ static NSLineBreakMode RCTNSLineBreakModeFromEllipsizeMode(EllipsizeMode ellipsi
-                   CGSize attachmentSize = attachment.bounds.size;
-                   CGRect glyphRect = [layoutManager boundingRectForGlyphRange:range inTextContainer:textContainer];
- 
--                  CGRect frame;
--                  CGFloat baseline = [layoutManager locationForGlyphAtIndex:range.location].y;
--
--                  frame = {{glyphRect.origin.x, glyphRect.origin.y + baseline - attachmentSize.height}, attachmentSize};
-+		  UIFont *font = [[textStorage attributedSubstringFromRange:range]
-+                           attribute:NSFontAttributeName
-+                             atIndex:0
-+                      effectiveRange:nil];
-+                  CGRect frame = {{glyphRect.origin.x, glyphRect.origin.y + glyphRect.size.height - attachmentSize.height + font.descender}, attachmentSize};
- 
-                   auto rect = facebook::react::Rect{
-                       facebook::react::Point{frame.origin.x, frame.origin.y},
