+++ shared/util/custom-emoji.native.tsx
@@ -1,14 +1,13 @@
 import type {Props} from './custom-emoji'
 import Image2 from '@/common-adapters/image2'
-import * as Styles from '@/styles'
 
 const CustomEmoji = (props: Props) => {
-  const {size, src, style} = props
+  const {size, src} = props
   const dimensions = {
     height: size,
-    transform: [{translateY: Styles.isAndroid ? 4 : 2}],
+    transform: [{translateY: 6}],
     width: size,
-    ...style,
+    ...props.style,
   }
 
   return <Image2 key={size} src={src} style={dimensions} />
