+++ shared/whats-new/versions.tsx
@@ -62,11 +62,7 @@ export const Last = ({seen}: VersionProps) => {
           </Kb.Box2>
           <Kb.Text type="BodySmall" allowFontScaling={true}>
             Keybase now supports custom emoji{' '}
-            <Kb.NativeEmoji
-              allowFontScaling={true}
-              size={Kb.Styles.globalMargins.small}
-              emojiName=":sparkles:"
-            />
+            <Kb.Emoji allowFontScaling={true} size={Kb.Styles.globalMargins.small} emojiName=":sparkles:" />
           </Kb.Text>
         </Kb.Box2>
       </NewFeatureRow>
@@ -80,7 +76,7 @@ export const LastLast = ({seen, onNavigate, onNavigateExternal}: VersionProps) =
       <VersionTitle title="Previous releases" />
       <NewFeatureRow image="release-5.3.0-ipad" noSeparator={true} seen={seen}>
         Keybase for iPad is here!{' '}
-        <Kb.NativeEmoji allowFontScaling={true} size={Kb.Styles.globalMargins.small} emojiName=":sparkles:" />{' '}
+        <Kb.Emoji allowFontScaling={true} size={Kb.Styles.globalMargins.small} emojiName=":sparkles:" />{' '}
         Download it from the App Store.
       </NewFeatureRow>
       <NewFeatureRow image="release-5.3.0-open-teams" seen={seen}>
@@ -97,7 +93,7 @@ export const LastLast = ({seen, onNavigate, onNavigateExternal}: VersionProps) =
         seen={seen}
       >
         Encrypt, decrypt, sign, and verify all from within Keybase.{' '}
-        <Kb.NativeEmoji
+        <Kb.Emoji
           allowFontScaling={true}
           size={Kb.Styles.globalMargins.small}
           emojiName=":closed_lock_with_key:"
@@ -151,12 +147,12 @@ export const LastLast = ({seen, onNavigate, onNavigateExternal}: VersionProps) =
       >
         You can now start a conversation with a phone number or email address.
         {` `}
-        <Kb.NativeEmoji allowFontScaling={true} size={Kb.Styles.globalMargins.small} emojiName=":phone:" />
+        <Kb.Emoji allowFontScaling={true} size={Kb.Styles.globalMargins.small} emojiName=":phone:" />
       </NewFeatureRow>
       <NewFeatureRow seen={seen} image={'release-4.7.0-pinned-messages'}>
         Chat admins can now pin messages.
         {` `}
-        <Kb.NativeEmoji size={Kb.Styles.globalMargins.small} emojiName=":pushpin:" />
+        <Kb.Emoji size={Kb.Styles.globalMargins.small} emojiName=":pushpin:" />
       </NewFeatureRow>
       <NewFeatureRow seen={seen} image={'release-4.7.0-keybase-fm'}>
         Listen to
